Don't process a dependency if it's met already.
Let's say that gem P depends on gem Q >= 1.0. Gem Q 1.0 is installed, but Q 2.0
has been released and depends on another gem, R. Installing P will (correctly)
result in Q 2.0 not being installed, since Q 1.0 meets the requirement. But Q
2.0's dependencies are erroneously considered, and we get an unnecessary
installation of R.

This fixes that bug so that if Q is already installed, the dependencies of its
latest version are not considered.